Move the data structures in a separate module
